Adds limit and offset implementation
Need this implemented to allow batch fetching for large entities volume

EntityQueryService has limit and offset specified in the proto as part
of the Query message. However, there's no implementation and it's has been
faux attributes.

This adds the implementation of limit and offset with the default
data storage ordering system.